refactor: re-configure the directory setting
