+++ src/(dashboard)/userAssets/components/assets-nft.tsx
@@ -1,0 +1,3 @@
+export const AssetsNft = () => {
+  return <div>Assets NFT</div>;
+};
