+++ src/(dashboard)/userAssets/components/assets-sft.tsx
@@ -1,0 +1,3 @@
+export const AssetsSft = () => {
+  return <div>Assets Sft</div>;
+};
